♻️ refactor(checkForString): 파일명 변경 및 적용
