+++ src/utils/checkForString.js
